+++ tests/test_visitors/test_ast/test_classes/test_base_classes/test_class_definition_arguments.py
@@ -3,7 +3,9 @@ import pytest
 from wemake_python_styleguide.violations.best_practices import (
     KwargsUnpackingInClassDefinitionViolation,
 )
-from wemake_python_styleguide.visitors.ast.classes import WrongClassDefVisitor
+from wemake_python_styleguide.visitors.ast.classes.classdef import (
+    WrongClassDefVisitor,
+)
 
 # Wrong: