+++ tests/test_visitors/test_ast/test_classes/test_wrong_class_body.py
@@ -3,7 +3,9 @@ import pytest
 from wemake_python_styleguide.violations.oop import (
     WrongClassBodyContentViolation,
 )
-from wemake_python_styleguide.visitors.ast.classes import WrongClassBodyVisitor
+from wemake_python_styleguide.visitors.ast.classes.classdef import (
+    WrongClassBodyVisitor,
+)
 
 class_body_template = """
 class ClassWithBody: