+++ CHANGELOG.md
@@ -19,9 +19,10 @@ Semantic versioning in our case means:
 
 ## 1.1.0 WIP
 
-### Features
+### Rules
 
 - Forbids to use `await` in `for` loop, #1600
+- Adds `WPS477`: do not allow `TypeVarTuple` after a `TypeVar` with a default #3265
 
 ### Command line utility